Factor image generation into a binary
This leaves the generation algorithm as a library that can be called
compiled to WASM. The output medium is now hidden behind a trait that
just knows how to set a pixel at a particular position to a given
`CellValue`. (What the calling program does with the eventual output is
up to the calling program).